backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.25.0
supabase>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
